+++ src/main/java/com/mordenkainen/equivalentenergistics/integration/ae2/storagechannel/IAEEMCStack.java
@@ -4,6 +4,6 @@ import appeng.api.storage.data.IAEStack;
 
 public interface IAEEMCStack extends IAEStack<IAEEMCStack> {
 
-    double getEMCValue();
+    long getEMCValue();
     
 }
